requirements.txt
pyPDF2
tabula-py==2.9.0
pandas
matplotlib
plotly
seaborn
nbformat
millify
streamlit
python-dotenv
tiktoken==0.4.0
faiss-cpu==1.7.4
protobuf
google-generativeai
langchain
langchain_core
langchain_experimental
langchain_google_genai
langchain_community
langchain_openai
fastapi
uvicorn
pydantic
requests